requirements/base.txt
pyautogui          # скриншоты, мышь, клавиатура (все ОС)
Pillow             # работа с изображениями
pytesseract        # OCR через Tesseract
